Extract MidiOpCode enum into controllers/midi/midiopcode.h
Part of a broader effort to clean up include dependencies as a step
towards a modular build.

Moved the MidiOpCode enum class and its operator<< and qHash declarations
from controllers/midi/midimessage.h into a new thin header
controllers/midi/midiopcode.h, with the implementations in
midiopcode.cpp. midimessage.h now includes midiopcode.h so all existing
code continues to work unchanged.

Updated control/controlbehavior.h and control/controlobject.h to include
midiopcode.h directly instead of the full midimessage.h. This breaks the
dependency from the control module on the controllers module, which was
the only reason control headers were pulling in Qt::Qml transitively.
