+++ src/controllers/midi/midimessage.cpp
@@ -1,14 +1,5 @@
 #include "controllers/midi/midiutils.h"
 
-QDebug operator<<(QDebug debug, MidiOpCode midiOpCode) {
-    debug << static_cast<uint8_t>(midiOpCode);
-    return debug;
-}
-
-qhash_seed_t qHash(MidiOpCode key, qhash_seed_t seed) {
-    return qHash(static_cast<uint8_t>(key), seed);
-}
-
 MidiKey::MidiKey()
         : status(0),
           control(0) {
